+++ pkg/morph/event/container/put_notary.go
@@ -4,53 +4,9 @@ import (
 	"fmt"
 
 	"github.com/nspcc-dev/neo-go/pkg/network/payload"
-	"github.com/nspcc-dev/neo-go/pkg/vm/opcode"
+	"github.com/nspcc-dev/neo-go/pkg/smartcontract/scparser"
 	"github.com/nspcc-dev/neofs-node/pkg/morph/event"
 )
-
-func (p *Put) setRawContainer(v []byte) {
-	if v != nil {
-		p.rawContainer = v
-	}
-}
-
-func (p *Put) setSignature(v []byte) {
-	if v != nil {
-		p.signature = v
-	}
-}
-
-func (p *Put) setPublicKey(v []byte) {
-	if v != nil {
-		p.publicKey = v
-	}
-}
-
-func (p *Put) setToken(v []byte) {
-	if v != nil {
-		p.token = v
-	}
-}
-
-func (p *Put) setName(v string) {
-	p.name = v
-}
-
-func (p *Put) setZone(v string) {
-	p.zone = v
-}
-
-func (p *Put) setMetaOnChain(v bool) {
-	p.metaOnChain = v
-}
-
-var putFieldSetters = []func(*Put, []byte){
-	// order on stack is reversed
-	(*Put).setToken,
-	(*Put).setPublicKey,
-	(*Put).setSignature,
-	(*Put).setRawContainer,
-}
 
 const (
 	// PutNotaryEvent is method name for container put operations
@@ -62,57 +18,66 @@ const (
 	PutNamedNotaryEvent = "putNamed"
 )
 
-func parsePutNotary(ev *Put, raw *payload.P2PNotaryRequest, ops []event.Op) error {
-	var (
-		currentOp opcode.Opcode
-		fieldNum  = 0
-	)
+const (
+	putArgCnt      = 4
+	putNamedArgCnt = 6 // `putNamed` has the same args as `put` + (name, zone) (2)
+)
 
-	switch len(ops) {
-	case expectedItemNumPut + 3:
-		enableMeta, err := event.BoolFromOpcode(ops[0])
-		if err != nil {
-			return fmt.Errorf("parse arg meta: %w", err)
-		}
-		ev.setMetaOnChain(enableMeta)
-
-		ops = ops[1:]
-
-		err = parseNamedArgs(ev, ops)
+func parsePutNotary(ev *Put, raw *payload.P2PNotaryRequest, args []scparser.PushedItem, t event.NotaryType) error {
+	switch l := len(args); l {
+	case putArgCnt + 3:
+		err := parseNamedArgs(ev, args[putArgCnt:])
 		if err != nil {
 			return err
 		}
-
-		ops = ops[2:]
-	case expectedItemNumPut + 2:
-		err := parseNamedArgs(ev, ops)
+		ev.metaOnChain, err = event.GetValueFromArg(args, l-1, t.String(), scparser.GetBoolFromInstr)
 		if err != nil {
 			return err
 		}
-
-		ops = ops[2:]
-	case expectedItemNumPut:
+	case putArgCnt + 2:
+		err := parseNamedArgs(ev, args[putArgCnt:])
+		if err != nil {
+			return err
+		}
+	case putArgCnt:
 	default:
-		return fmt.Errorf("unknown number of args: %d", len(ops))
+		return fmt.Errorf("%s: unknown number of args: %d", t, l)
 	}
 
-	for _, op := range ops {
-		currentOp = op.Code()
-
-		switch {
-		case opcode.PUSHDATA1 <= currentOp && currentOp <= opcode.PUSHDATA4:
-			if fieldNum == expectedItemNumPut {
-				return event.UnexpectedArgNumErr(PutNotaryEvent)
-			}
-
-			putFieldSetters[fieldNum](ev, op.Param())
-			fieldNum++
-		default:
-			return event.UnexpectedOpcode(PutNotaryEvent, op.Code())
-		}
+	var err error
+	ev.rawContainer, err = event.GetValueFromArg(args, 0, t.String(), scparser.GetBytesFromInstr)
+	if err != nil {
+		return err
+	}
+	ev.signature, err = event.GetValueFromArg(args, 1, t.String(), scparser.GetBytesFromInstr)
+	if err != nil {
+		return err
+	}
+	ev.publicKey, err = event.GetValueFromArg(args, 2, t.String(), scparser.GetBytesFromInstr)
+	if err != nil {
+		return err
+	}
+	ev.token, err = event.GetValueFromArg(args, 3, t.String(), scparser.GetBytesFromInstr)
+	if err != nil {
+		return err
 	}
 
 	ev.notaryRequest = raw
+
+	return nil
+}
+
+func parseNamedArgs(ev *Put, args []scparser.PushedItem) error {
+	var err error
+	ev.name, err = scparser.GetStringFromInstr(args[0].Instruction)
+	if err != nil {
+		return fmt.Errorf("zone: %w", err)
+	}
+
+	ev.zone, err = scparser.GetStringFromInstr(args[1].Instruction)
+	if err != nil {
+		return fmt.Errorf("name: %w", err)
+	}
 
 	return nil
 }
@@ -121,7 +86,7 @@ func parsePutNotary(ev *Put, raw *payload.P2PNotaryRequest, ops []event.Op) erro
 func ParsePutNotary(ne event.NotaryEvent) (event.Event, error) {
 	var ev Put
 
-	err := parsePutNotary(&ev, ne.Raw(), ne.Params())
+	err := parsePutNotary(&ev, ne.Raw(), ne.Params(), ne.Type())
 	if err != nil {
 		return nil, err
 	}
@@ -131,50 +96,16 @@ func ParsePutNotary(ne event.NotaryEvent) (event.Event, error) {
 
 // ParsePutNamedNotary parses PutNamed event structure from generic event.NotaryEvent.
 func ParsePutNamedNotary(ne event.NotaryEvent) (event.Event, error) {
-	ops := ne.Params()
-
-	const putNamedAdditionalArgs = 2 // PutNamed has same args as Put + (name, zone) (2)
-
-	if len(ops) != expectedItemNumPut+putNamedAdditionalArgs {
-		return nil, event.UnexpectedArgNumErr(PutNamedNotaryEvent)
-	}
-
-	var (
-		ev  PutNamed
-		err error
-	)
-
-	err = parseNamedArgs(&ev, ops)
+	args, err := event.GetArgs(ne, putNamedArgCnt)
 	if err != nil {
 		return nil, err
 	}
 
-	err = parsePutNotary(&ev.Put, ne.Raw(), ops[putNamedAdditionalArgs:])
+	var ev Put
+	err = parsePutNotary(&ev, ne.Raw(), args, ne.Type())
 	if err != nil {
 		return nil, err
 	}
 
 	return ev, nil
-}
-
-type putEvNamed interface {
-	setName(v string)
-	setZone(v string)
-}
-
-func parseNamedArgs(p putEvNamed, ops []event.Op) error {
-	zone, err := event.StringFromOpcode(ops[0])
-	if err != nil {
-		return fmt.Errorf("parse arg zone: %w", err)
-	}
-
-	name, err := event.StringFromOpcode(ops[1])
-	if err != nil {
-		return fmt.Errorf("parse arg name: %w", err)
-	}
-
-	p.setZone(zone)
-	p.setName(name)
-
-	return nil
 }